+++ packages/wallet/src/W3mFrameProvider.ts
@@ -135,6 +135,10 @@ export class W3mFrameProvider {
     return W3mFrameStorage.get(W3mFrameConstants.EMAIL)
   }
 
+  public rejectRpcRequest() {
+    this.rpcRequestResolver?.reject()
+  }
+
   public async connectEmail(payload: W3mFrameTypes.Requests['AppConnectEmailRequest']) {
     await this.w3mFrame.frameLoadPromise
     W3mFrameHelpers.checkIfAllowedToTriggerEmail()