+++ examples/next-wagmi/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @examples/next-wagmi
 
+## 4.0.9
+
+### Patch Changes
+
+- Add all rpc methods + auto reject when modal closes
+
+- Updated dependencies []:
+  - @web3modal/wagmi@4.0.9
+
 ## 4.0.8
 
 ### Patch Changes