+++ packages/wallet/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @web3modal/wallet
 
+## 4.0.9
+
+### Patch Changes
+
+- Add all rpc methods + auto reject when modal closes
+
 ## 4.0.8
 
 ### Patch Changes